feat: add download api key page
