+++ src/app/view/download-api-key/download-api-key.component.html
@@ -1,0 +1,10 @@
+<div
+  class="container flex flex-column align-items-center justify-content-center h-full w-full"
+>
+  <div class="text-center">
+    <h1>Download API Key</h1>
+    <button class="btn btn-primary" (click)="createAccessKey()">
+      Download API Key
+    </button>
+  </div>
+</div>
